+++ src/ResourceManager/DataMigration/Commands.DataMigration/Cmdlets/GetProjectCmdlet.cs
@@ -24,7 +24,7 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
     /// <summary>
     /// Class for the command let that creates a new instance of the Data Migration Service.
     /// </summary>
-    [Cmdlet(VerbsCommon.Get, "AzureRmDataMigrationProject", DefaultParameterSetName = ComponentNameParameterSet), OutputType(typeof(PSProject))]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataMigrationProject", DefaultParameterSetName = ComponentNameParameterSet), OutputType(typeof(PSProject))]
     [Alias("Get-AzureRmDmsProject")]
     public class GetProjectCmdlet : DataMigrationCmdlet
     {