+++ src/ResourceManager/ServiceBus/Commands.ServiceBus/Cmdlets/MigrationConfiguration/StartAzureServiceBusMigrationConfiguration.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Migration
     /// <summary>
     /// 'Start-AzureRmServicebusMigration' Cmdlet Creates an new Migration configuration of Standard to Premium
     /// </summary>
-    [Cmdlet(VerbsLifecycle.Start, ServicebusMigrationConfigurationVerb, DefaultParameterSetName = MigrationConfigurationParameterSet, SupportsShouldProcess = true), OutputType(typeof(PSServiceBusDRConfigurationAttributes))]
+    [Cmdlet("Start", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ServiceBusMigration", DefaultParameterSetName = MigrationConfigurationParameterSet, SupportsShouldProcess = true), OutputType(typeof(PSServiceBusDRConfigurationAttributes))]
     public class StartAzureServiceBusMigrationConfiguration : AzureServiceBusCmdletBase
     {
         [Parameter(Mandatory = true, Position = 0, HelpMessage = "Resource Group Name")]